api-analise-imagens/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Async Tasks
celery==5.3.6
redis==5.0.3
flower==2.0.1

# OCR e Visão Computacional
easyocr==1.7.0
opencv-python-headless==4.9.0.80
Pillow<10.0.0
numpy==1.26.3

# Storage
minio==7.2.3

# LLM APIs
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2

# Utilitários
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
rapidfuzz==3.6.1
python-dotenv==1.0.0

# Logging e Monitoramento
structlog==24.1.0

# Testes
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Qualidade de Código
ruff==0.1.14
mypy==1.8.0
black==24.1.1
